requirements.txt
openai
python-dotenv
discord.py
faster-whisper
